+++ tests/Feature/Api/PetTest.php
@@ -365,116 +365,4 @@ class PetTest extends TestCase
             'uuid' => $pet->uuid,
         ]);
     }
-
-    final public function testAttachPetsGroupsApi()
-    {
-        $user = factory(User::class)->create();
-        $group = factory(Group::class)->create();
-        $pet = factory(Pet::class)->create();
-        $pet->user_uuid = $user->uuid;
-        $pet->save();
-        $group->user_uuid = $user->uuid;
-        $group->save();
-        $group->users()->attach($user, [
-            'is_admin' => true,
-            'status' => Group::ACCEPTED_STATUS
-        ]);
-        $response = $this->actingAs($user, 'api')->json(
-            'POST',
-            '/api/pets/' . $pet->uuid . '/groups/' . $group->uuid . '/attach',
-            []
-        );
-        $response->assertOk();
-
-        $this->assertDatabaseHas('groupables', [
-            'groupable_type' => 'App\\Pet',
-            'groupable_id' => $pet->uuid,
-            'group_uuid' => $group->uuid,
-        ]);
-    }
-
-    final public function testDetachPetsGroupsApi()
-    {
-        $user = factory(User::class)->create();
-        $group = factory(Group::class)->create();
-        $pet = factory(Pet::class)->create();
-        $pet->user_uuid = $user->uuid;
-        $pet->save();
-        $group->user_uuid = $user->uuid;
-        $group->save();
-        $group->users()->attach($user, [
-            'is_admin' => true,
-            'status' => Group::ACCEPTED_STATUS
-        ]);
-        $pet->groups()->attach($group, [
-            'status' => Group::ACCEPTED_STATUS
-        ]);
-        $response = $this->actingAs($user, 'api')->json(
-            'POST',
-            '/api/pets/' . $pet->uuid . '/groups/' . $group->uuid . '/detach',
-            []
-        );
-        $response->assertOk();
-
-        $this->assertDeleted('groupables', [
-            'groupable_type' => 'App\\Pet',
-            'groupable_id' => $pet->uuid,
-            'group_uuid' => $group->uuid,
-        ]);
-    }
-
-    /**
-     * @test testUserCanViewThePetGroups
-     */
-    final public function testUserCanViewThePetGroups()
-    {
-        $user = factory(User::class)->create();
-        $pet = factory(Pet::class)->create();
-        $group = factory(Group::class)->create();
-        $group->user_uuid = $user->uuid;
-        $group->save();
-        $group->users()->attach($user, [
-            'is_admin' => true,
-            'status' => Group::ACCEPTED_STATUS
-        ]);
-        $pet->user_uuid = $user->uuid;
-        $pet->save();
-        $pet->groups()->attach($group, [
-            'status' => Group::ACCEPTED_STATUS
-        ]);
-        $this->assertDatabaseHas('groupables', [
-            'groupable_type' => 'App\\Pet',
-            'groupable_id' => $pet->uuid,
-            'group_uuid' => $group->uuid,
-        ]);
-        $response = $this->actingAs($user, 'api')->json(
-            'GET',
-            '/api/pets/' . $pet->uuid . '/groups',
-            []
-        );
-        $response->assertJsonStructure([
-            'data' => [
-                [
-                    'uuid',
-                    'user',
-                    'photo',
-                    'description',
-                    'is_public'
-                ]
-            ]
-        ]);
-        // Assert User UUID
-        $response->assertJsonFragment([
-            'uuid' => $user->uuid,
-        ]);
-        // Assert Group UUID
-        $response->assertJsonFragment([
-            'uuid' => $group->uuid,
-        ]);
-        // Assert Photo UUID
-        $response->assertJsonFragment([
-            'uuid' => json_decode($response->getContent())->data[0]->photo->uuid,
-        ]);
-        $response->assertOk();
-    }
 }
