+++ tests/Feature/Stories/UserCanDetachPetOfGroupTest.php
@@ -1,0 +1,46 @@
+<?php
+
+
+namespace Tests\Feature\Stories;
+
+
+use App\Group;
+use App\Pet;
+use App\User;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+use Tests\TestCase;
+
+class UserCanDetachPetOfGroupTest extends TestCase
+{
+    use RefreshDatabase;
+
+    final public function testDetachPetsGroupsApi()
+    {
+        $user = factory(User::class)->create();
+        $group = factory(Group::class)->create();
+        $pet = factory(Pet::class)->create();
+        $pet->user_uuid = $user->uuid;
+        $pet->save();
+        $group->user_uuid = $user->uuid;
+        $group->save();
+        $group->users()->attach($user, [
+            'is_admin' => true,
+            'status' => Group::ACCEPTED_STATUS
+        ]);
+        $pet->groups()->attach($group, [
+            'status' => Group::ACCEPTED_STATUS
+        ]);
+        $response = $this->actingAs($user, 'api')->json(
+            'POST',
+            '/api/pets/' . $pet->uuid . '/groups/' . $group->uuid . '/detach',
+            []
+        );
+        $response->assertOk();
+
+        $this->assertDeleted('groupables', [
+            'groupable_type' => 'App\\Pet',
+            'groupable_id' => $pet->uuid,
+            'group_uuid' => $group->uuid,
+        ]);
+    }
+}
